moved gnps downloader to gnps folder
